backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
langchain==0.1.0
langchain-google-genai==0.0.12
langchain-community==0.0.20
langchain-text-splitters==0.0.1
langchain-core==0.1.32
google-generativeai==0.3.0
pymongo==4.6.0
chromadb==0.4.14
easyocr==1.7.0
pdfplumber==0.10.3
pillow==10.0.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.1
aiofiles==23.2.1
reportlab==4.0.7
